commit4

+++ keep-note.sql
@@ -9,7 +9,7 @@ CREATE TABLE Category (
     category_id INT PRIMARY KEY,
     category_name VARCHAR(20),
     category_descr VARCHAR(20),
-    category_creation_date DATE,
+    category_creation_date DATETIME,
     category_creator VARCHAR(20)
 );
 CREATE TABLE Reminder (
@@ -175,25 +175,14 @@ delete from UserNote where note_id=2;
 
 delete from Note where note_id=2;
 
-
 DELIMITER //
-create trigger del_note before delete on Note FOR EACH ROW Begin delete from UserNote where note_id= old.note_id ;
-DELETE FROM NoteReminder WHERE note_id = old.note_id;
-DELETE FROM NoteCategory WHERE note_id = old.note_id; 
-END;
- 
-DELIMITER // 
-delete from Note where note_id=3;  
-
+create trigger del_note before delete on Note FOR EACH ROW Begin delete from UserNote where note_id=old.note_id; delete from NoteReminder where note_id=old.note_id; delete from NoteCategory where note_id=old.note_id; END; //
+DELIMITER ;
+delete from Note where note_id=3;
 DELIMITER //
-create trigger del_user before delete on User FOR EACH ROW Begin delete from UserNote where note_id=old.user_id;
- delete from NoteReminder where note_id=old.user_id; 
- delete from NoteCategory where note_id=old.user_id;
-END;
-
-DELIMITER //
-delete from User where user_id="1";  
-
+create trigger del_user before delete on User FOR EACH ROW Begin delete from UserNote where note_id=old.user_id; delete from NoteReminder where note_id=old.user_id; delete from NoteCategory where note_id=old.user_id; END; //
+DELIMITER ;
+delete from User where user_id="1";
